Use a shared package in RTDEWriter
As we have to send a full RTDE package each time we send data over RTDE,
but there are fields that are not masked, we have to know the state of our
data package each time we send a new data package.

This commit adds a buffered package to the RTDEWriter that will hold the state
of the full data package. After sending masked values, the corresponding mask
will be set to 0 after sending.

In order not to loose thread safety, a lock_guard is added to each send function.
This way, only one send function can be active at a single point in time.

The actual sending mechanism is still decoupled from the send methods.