dashboard_client: Allow a larger timeout on initial connection
Apparently, if we connect to the socket right after it became available,
it can happen, that we don't get a "connected" message from the dashboard
server in a second.

This commit should not only increase the timeout on the initial read to a
value that should be sufficient, but also establishes a retry with a warning
output.