Added the new primary client. The primary client can through its consumer get the message handlers and from the message handlers get the latest message of each type.
The primary client checks whether the robot is in remote control or not, as it should not be able to send script if the robot is not in remote control. This is not true if the robot is simulated. Therefore, when constructing a primary client instance it is needed to specify if the robot is simulated or real. I was not able to find a way for the primary client to check whether it was real or simulated automatically.

The primary client now automatically checks whether the checksum is correct when starting.