Add a cmake option to activate address sanitizers (#146)
* Add a cmake option to activate address sanitizers

* Fix unique_ptr type for comm buffer to array

We keep a pointer to the raw data buffer received from the comm interfaces
in our packages. This was held as a unique_ptr<uint8_t> while we were storing
a uint8_t type inside. This commit changes the pointer's type to the correct
one.

* Fix memory leak in tests

test_tcp_server had a memory leak by not deleting manually allocated clients.