Changed spline interpolation to use the last commanded joint velocity when calculating the next spline in the trajectory
Also limited the acceleration to the maximum joint acceleration of the target accelerations calculated from the spline. But still added a constant acceleration when slowing down is active, as the acceleration in the trajectory no longer can be used.

This will fix the issue with discontinuous motion when speed scaling is active