Created new exception MissingArgument
After further thought, I dont think the Incompatible version exception is appropriate for one of the exception calls, as the robot is not actually incompatible, the function just needs more information. The new exception conveys that, i think. I can't quite make my mind up about whether it should mention that this is due to the robot version, though, so I have left it out for now.