CI: Fix flaky example runs (#223)
The example runs seem to fail quite often. This PR aims at fixing that.

It seems that the simulated robot is reporting readiness for executing programs
a bit too early and if we execute the program rightaway it is not started correctly.
This only seems to happen in CI though.