Adapt RTDE output recipe based on robot response (#221)
Not all RTDE outputs are available in all SW versions. When working with
a fleet of robots with heterogeneous SW versions it's not practical to
have multiple RTDE output recipes to account for this.
This PR aims at solving this problem by adapting the RTDE output recipe
on-the-fly based on the robot response to the RTDE request. Outputs
which are not available for this particular SW version are simply
removed from the output recipe and a trimmed down output recipe is sent
to the robot.

Potential issues / improvements:
- variables removed from the output recipe only cause a warning. When
these variables are "optional" this is a fine behaviour but it could
potentially be masking incorrect variable names. Hence, this is an opt-in feature.
- an alternate implementation could be to hardcode the compatibility
information for RTDE outputs in `data_package.cpp` and validate the
output recipe beforehand. This seems less flexible though.
- the same logic could potentially be applied to the input recipe.
Having optional inputs seems less likely though.