Add instruction executor for high-level robot control (#242)
With the trajectory point interface there is a method of executing
motions with only the urcl present as shown in
https://github.com/UniversalRobots/Universal_Robots_Client_Library/blob/master/examples/trajectory_point_interface.cpp

However, usage is a bit cumbersome, we might want to support a one-liner
for executing ptp motions, instead.

This PR adds the `InstructionExecutor`, a high-level module that uses
existing functionality to provide simple-to-use interfaces, currently to
execute MoveJ, MoveL and sequences of the two. In future, we might
extend that to other instructions such as other motion types, but
possibly also other functionality, hence the more general name.