Drop checking programState when loading a program
When loading a program from an absolute path or a subdirectory instead
of a file being directly in /programs (or /ursim/programs in the case of
URSim) the check "programState" will only return the filename without
the path while we compare it against the path that was requested to
load.

However, that second check should not be necessary, as the "load" call
will already make sure that the program is loading correctly.