Re-add wait for stopped programState
We load the program, which will fail if the program is not found or the
requested file does  not contain a valid program. Afterwards, we wait
until the program state is stopped with a  program named the same as the
requested program. We cannot check the full file path here, but  the
important thing is that the program state is stopped.