Fix DashboardClient load program from subdir (#269)
When loading a program from an absolute path or a subdirectory instead
of a file being directly in /programs (or /ursim/programs in the case of
URSim) the check "programState" will only return the filename without
the path while we compare it against the path that was requested to
load.
    
This commit only compares the actual program's filename, as the full path
is already checked during the load operation.