In servo mode always allow targets close to current pose (#273)
If the target has been interpolating too far away from the robot, so it
would trigger the limit check, but the target is close to where the
robot currently is, also accept the command. This can, for example,
happen if a command series was too fast for the robot to follow, which
triggers a path deviation on the commanding side and that is mitigated
by a hold-position command.