Try catch RTDE setup (#285)
In #266, we improved the behavior of the urcl regarding reconnection
attempts and timeouts for sockets. However, there are still a few
problems (more to come later). I propose to fix one of them here to keep
the scope of the PRs small.

`RTDEClient::setupCommunication` is called in a while loop, and inside
`RTDEClient::setupCommunication` there is a call to
`pipeline_->init(...)` which in turn calls
`producer_.setupProducer(...)`. `setupProducer` (see
[here](https://github.com/UniversalRobots/Universal_Robots_Client_Library/blob/62d34dff101bc460a4d36692405fd076974b8791/include/ur_client_library/comm/producer.h#L67))
throws if the connection of the `stream_` (which is just a tcp socket in
my understanding) fails. Hence, if the RTDE sockets are not available,
this causes the initialization sequence to terminate immediately and not
retry for the desired `max_initialization_attempts`.

Now that I have a deeper understanding of the URCL after digging deeper
a bit (sorry that I didn't bring that up earlier), I don't fully agree
with the `rtde_initialization_attempts` and
`rtde_initialization_timeout` anymore since the RTDE sockets are just
usual tcp sockets that are already subject to
`socket_reconnect_attempts` and `socket_reconnection_timeout`. This
means that currently it's just a while loop of connecting tcp sockets in
a while loop for the RTDE connection setup which is maybe not really
necessary and just creates a lot of delays. But it's also possible that
I miss an important detail about RTDE comm so maybe my feeling is wrong
here.