[start_ursim.sh] Use direct web pages instead of GitHub API to download URCap (#308)
The API is rate-restricted on an IP address level. And with IPv4 that IP
address is potentially shared between many users of the same ISP.

This PR uses the release web pages to find the latest release.
Basically, it uses the link to the latest release which will be
forwarded by GitHub to a web page using the latest version. Thus, we can
get the latest tag from the redirected URL and use that to build a valid
download link for the latest release.