Fix robot message type POPUP (#335)
Before it was using the name PROGRAM_STATE_MESSAGE_VARIABLE_UPDATE

As far as I can see that error has been there for all eternity, but that
type es never used anywhere.

Strictly speaking, this would be breaking API, but the old naming was
clearly wrong, so users should not rely on that faulty behavior. Also
parsing that type is not implemented in the library, so this is most
probably not used by anybody. That would be my argument to make that
change without major-bumping the library. @urrsk @urmahp would you
agree?