Add ScriptReader for script template parsing (#343)
This PR adds a script parser that extends the templating capabilities
when loading script code.

In particular that allows
- Variable substitution (exists already, but this was refactored)
- Conditionals
- Includes

With this we can add script code that only compiles on later software
versions only if the required version is used. This way, older software
versions can still be used, the new feature will simply not be present
there. So far, when using new script code functions, the code would not
have compiled on newer versions effectively increasing the minimal
required software version.

---------

Co-authored-by: Mads Holm Peters <79145214+urmahp@users.noreply.github.com>