Install endian.h and add that to the target include directories...
... on Windows and MacOS

When installing this library the source directory might not be available,
thus we need to install the header. However, we need to make sure it is
doesn't spill into the global include directories to shadow and
os-specific installations e.g. on Linux.

Installing it to a 3rdparty folder and adding that to the target_include_directories
on Windows should do the trick from my understanding.