Stop torque_control example after timeout
On a simulated robot, torque_control commands don't have any effect. As
a result, the torque_control example runs forever when being run on a
simulated robot.
This commit makes it use the second argument as a timeout after which
the example is stopped. If the timeout is hit, it will print a warning
stating that this is expected on a simulated robot.

This should make the behavior similar to the other examples and make the
CI run with that example.