Succeed power on command on PolyScope 5 when robot is running (#397)
We check for the robot state being IDLE and retry that with a very high
timeout by default. This commit also accepts the robot to be in RUNNING
in order to succeed.

Otherwise the robot will reply "Powering on" through the dashboard
interface, but the state will stay unchanged because it is already
idling. Thus, the call will fail (and block very long) while the pure
dashboard call would be succeeding.