Do not disconnect dashboard_client on read timeout (#414)
Currently, we disconnect from the dashboard server whenever we don't
receive an answer inside the configured timeout and we throw an
exception. The idea behind that was, that if we don't get an answer in
time there's probably something wrong with the connection, anyway.
However, it could just be that the command in question takes longer to
process. There are a couple of commands where we manually add higher
timeouts, such as generating the flight report, but also powering on the
robot after an update can take longer if the firmware has to be updated.

With the change in this commit, we still throw the exception leaving it
up to the application layer to handle this. However, keeping the
connection alive allows running other dashboard commands to inspect
what's going on or simply retry with another timeout setting.