Avoid allocating memory during parsing
Creating new objects on the heap isn't really RT-capable and
undeterministic. This adds the possibility to parse incoming data
packages into a pre-allocated package to avoid having to create new
objects on the fly.

To achieve this, the pipeline in the RTDE client has been replaced by a
double buffer with two pre-allocated objects.